Link attributes
I added a color for links to the stylesheet so it would stop defaulting to the foundation file for the color. I made the links the same blue color #0055a5 and made the text heavier to further distinguish it from other text for color-blind or visually impaired users. I also added underline on hover.